Fix padding issues in compose layouts when IME is visible
A workaround to resize a view layout when IME is open was unintentionally applied in compose layouts which resulted in too much padding. As the workaround is only needed in the widget config layouts it is moved to the AbstractWidgetConfigActivity.
